## 4.1.1.0
* Video library update
* Video list dialog can be toggled to show URLs
* Add Edit Mode for Vehicle Models
* Add deprecation throughout codebase with target version for removal
* Add retry attempts for failed tile downloads during bulk map download
* Source code updates to make Java more portable
* Integrate filesystem abstraction layer
* Various bug fixes
* New gradle plugin to support ATAK plugin development

+++ atak/ATAKMapEngine/lib/src/main/java/com/atakmap/opengl/GLTextureAtlas.java
@@ -15,8 +15,6 @@ import android.graphics.Canvas;
 import android.graphics.RectF;
 import android.opengl.GLUtils;
 
-import com.atakmap.coremap.log.Log;
-
 public class GLTextureAtlas {
 
     private final static Comparator<Rect> HORIZONTAL_FREE_COMPARATOR = new Comparator<Rect>() {
@@ -220,7 +218,7 @@ public class GLTextureAtlas {
         if (this.fixedIconSize) {
             final int index = this.getIndex(key);
             final int numIconCols = (this.texSize / this.iconSize);
-            rect.top = (index / numIconCols) * this.iconSize;
+            rect.top = (index / (float)numIconCols) * this.iconSize;
             rect.left = (index % numIconCols) * this.iconSize;
             rect.bottom = rect.top + this.iconSize - 1;
             rect.right = rect.left + this.iconSize - 1;
